Make test data dynamic.
