--- packages/extension/src/view/devtools/stateProviders/filterManagementStore/utils/tests/testUtils.ts
@@ -1,401 +1,0 @@
-/*
- * Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-/*
- * Internal dependencies
- */
-import type { CookieTableData } from '../../../../cookies.types';
-
-export const frameFilteredCookies: { [key: string]: CookieTableData } = {
-  LSOLH: {
-    analytics: {
-      category: 'Functional',
-      dataController: 'Google',
-      description: 'This cookie is for authentication with your Google account',
-      domain: 'accounts.google.com',
-      gdprUrl: 'https://privacy.google.com/take-control.html',
-      name: 'LSOLH',
-      platform: 'Google',
-      retention: '1 year',
-      wildcard: '0',
-    },
-    frameIdList: [0],
-    headerType: 'request',
-    isFirstParty: false,
-    parsedCookie: {
-      domain: 'accounts.google.com',
-      expires: '2024-08-08T07:51:01.301Z',
-      httponly: false,
-      name: 'LSOLH',
-      path: '/',
-      samesite: 'none',
-      secure: true,
-      value:
-        '|_SVI_EMmpku6Lz4ADGBAiP01BRURIZl9xZWR4OHp5aUhCQ1h5RnVzcXFlb1NPUUVQeXVPZGlSTjBLeXdLMnRMOGpfd3JQZDk0STh3ZWJVNA_:|28192806:3705',
-    },
-    url: 'https://accounts.google.com/gsi/status?client_id=549796576131-8aoqmiheqf84a8bh3dl10ec2kp2v6e7p.apps.googleusercontent.com&as=ujHsjnl%2BJlozSZHE2Z6o2A',
-    isCookieSet: true,
-  },
-  NID: {
-    analytics: {
-      category: 'Marketing',
-      dataController: 'Google',
-      description:
-        'This cookies is used to collect website statistics and track conversion rates and Google ad personalisation',
-      domain: 'google.com',
-      gdprUrl: 'https://privacy.google.com/take-control.html',
-      name: 'NID',
-      platform: 'Google',
-      retention: '1 year',
-      wildcard: '0',
-    },
-    frameIdList: [3964, 3967, 0],
-    headerType: 'request',
-    isFirstParty: false,
-    parsedCookie: {
-      domain: '.google.com',
-      expires: '2024-02-16T08:07:17.302Z',
-      httponly: true,
-      name: 'NID',
-      path: '/',
-      samesite: 'none',
-      secure: true,
-      value:
-        '511=Wrres5z7l_wUKKUzHhf2KIqAnj6papQrcGn1F12gPxd2SdxEGS-567t-hT7Yuzb8tJsENk8pT_9xCJ1UMPLbXKQk4mPUCtMp7VeFXp0yi8t0xPMvqgfUKKBpQwV8XmrveNPPwLmwxriIpyicaBYoJ-ZzfHvqnCZDya9rU_s_JwGk5JvIDTqIEYVQKzKJDyFQjiAINLhRFCmWOTWF7X3C3mYKVaUQrT4JLhZRAGEoj4Qz_ig5MbxpLB372-PjFrNofa5FvfmYWn3uBEFYm0UvAqhhyOcsND5eAf_iYYWufJAVGReSyfWZ0spF25g4TS-pUdU8l3zDJC-7',
-    },
-    url: 'https://www.google.com/recaptcha/api2/anchor?ar=1&k=6Le6JpwaAAAAAFkASVap1OUThl-lQSJC0r9kLl2I&co=aHR0cHM6Ly9ydGNhbXAuY29tOjQ0Mw..&hl=en&v=QybaJej5brGL8d7EvWmfKMZU&theme=light&size=invisible&badge=bottomright&cb=hztbhk44za6a',
-    isCookieSet: true,
-  },
-  _GRECAPTCHA: {
-    analytics: {
-      category: 'Functional',
-      dataController: 'Google',
-      description:
-        'Google reCAPTCHA sets a necessary cookie (_GRECAPTCHA) when executed for the purpose of providing its risk analysis.',
-      domain: 'google.com',
-      gdprUrl: 'https://privacy.google.com/take-control.html',
-      name: '_GRECAPTCHA',
-      platform: 'Google reCAPTCHA',
-      retention: '179 days',
-      wildcard: '0',
-    },
-    frameIdList: [3964, 3967, 0],
-    headerType: 'request',
-    isFirstParty: false,
-    parsedCookie: {
-      domain: 'www.google.com',
-      expires: '2024-02-05T07:24:44.406Z',
-      httponly: true,
-      name: '_GRECAPTCHA',
-      path: '/recaptcha',
-      samesite: 'none',
-      secure: true,
-      value:
-        '09AD1dMiqkTKi6u46RctmyHstxVOcjBTlEqyJzgrgJ_8FrYPi53-AEJawWg0O3h_CJ2GvUO34mPGerSausVf-VAFY',
-    },
-    url: 'https://www.google.com/recaptcha/api2/anchor?ar=1&k=6Le6JpwaAAAAAFkASVap1OUThl-lQSJC0r9kLl2I&co=aHR0cHM6Ly9ydGNhbXAuY29tOjQ0Mw..&hl=en&v=QybaJej5brGL8d7EvWmfKMZU&theme=light&size=invisible&badge=bottomright&cb=hztbhk44za6a',
-    isCookieSet: true,
-  },
-  '__Host-3PLSID': {
-    analytics: {
-      category: 'Functional',
-      dataController: 'Google',
-      description: 'Used to sign in with Google account.',
-      domain: 'accounts.google.com',
-      gdprUrl: 'https://privacy.google.com/take-control.html',
-      name: '__Host-3PLSID',
-      platform: 'Google',
-      retention: '1 year',
-      wildcard: '0',
-    },
-    frameIdList: [0],
-    headerType: 'request',
-    isFirstParty: false,
-    parsedCookie: {
-      domain: 'accounts.google.com',
-      expires: '2024-09-12T07:41:15.169Z',
-      httponly: true,
-      name: '__Host-3PLSID',
-      path: '/',
-      samesite: 'none',
-      secure: true,
-      value:
-        'o.console.cloud.google.com|o.drive.google.com|o.gds.google.com|o.groups.google.com|o.mail.google.com|o.play.google.com|s.IN|s.youtube:ZwifqQn_cngJDleofZMlcFXPZ5i7Xbijxx8WIqjWTUwlcYblWDItPd1imNYpyLPgykageQ.',
-    },
-    url: 'https://accounts.google.com/gsi/status?client_id=549796576131-8aoqmiheqf84a8bh3dl10ec2kp2v6e7p.apps.googleusercontent.com&as=ujHsjnl%2BJlozSZHE2Z6o2A',
-    isCookieSet: true,
-  },
-  '__Secure-3PAPISID': {
-    analytics: {
-      category: 'Marketing',
-      dataController: 'Google',
-      description:
-        'Profiles the interests of website visitors to serve relevant and personalised ads through retargeting.',
-      domain: 'google.com',
-      gdprUrl: 'https://privacy.google.com/take-control.html',
-      name: '__Secure-3PAPISID',
-      platform: 'Google Ads',
-      retention: '2 years',
-      wildcard: '0',
-    },
-    frameIdList: [3964, 3967, 0],
-    headerType: 'request',
-    isFirstParty: false,
-    parsedCookie: {
-      domain: '.google.com',
-      expires: '2024-09-12T07:41:15.170Z',
-      httponly: false,
-      name: '__Secure-3PAPISID',
-      path: '/',
-      samesite: 'none',
-      secure: true,
-      value: 'c9qnCgR5IUVPfo1H/AIs-Kk2Nz05sIxysj',
-    },
-    url: 'https://www.google.com/recaptcha/api2/anchor?ar=1&k=6Le6JpwaAAAAAFkASVap1OUThl-lQSJC0r9kLl2I&co=aHR0cHM6Ly9ydGNhbXAuY29tOjQ0Mw..&hl=en&v=QybaJej5brGL8d7EvWmfKMZU&theme=light&size=invisible&badge=bottomright&cb=hztbhk44za6a',
-    isCookieSet: true,
-  },
-  '__Secure-3PSID': {
-    analytics: {
-      category: 'Marketing',
-      dataController: 'Google',
-      description:
-        'Targeting cookie. Used to profile the interests of website visitors and display relevant and personalised Google ads.',
-      domain: 'google.com',
-      gdprUrl: 'https://privacy.google.com/take-control.html',
-      name: '__Secure-3PSID',
-      platform: 'Google Ads',
-      retention: '2 years',
-      wildcard: '0',
-    },
-    frameIdList: [3964, 3967, 0],
-    headerType: 'request',
-    isFirstParty: false,
-    parsedCookie: {
-      domain: '.google.com',
-      expires: '2024-09-12T07:41:15.169Z',
-      httponly: true,
-      name: '__Secure-3PSID',
-      path: '/',
-      samesite: 'none',
-      secure: true,
-      value:
-        'ZwifqbzMsjJC1kLXz7i_uPqhDchs7aU9v-wDqBe7Cdl9jna1ugf8--pdLnyhOJ96RPM5vQ.',
-    },
-    url: 'https://www.google.com/recaptcha/api2/anchor?ar=1&k=6Le6JpwaAAAAAFkASVap1OUThl-lQSJC0r9kLl2I&co=aHR0cHM6Ly9ydGNhbXAuY29tOjQ0Mw..&hl=en&v=QybaJej5brGL8d7EvWmfKMZU&theme=light&size=invisible&badge=bottomright&cb=hztbhk44za6a',
-    isCookieSet: true,
-  },
-  '__Secure-3PSIDCC': {
-    analytics: {
-      category: 'Marketing',
-      dataController: 'Google',
-      description:
-        'Targeting cookie. Used to create a user profile and display relevant and personalised Google Ads to the user.',
-      domain: 'google.com',
-      gdprUrl: 'https://privacy.google.com/take-control.html',
-      name: '__Secure-3PSIDCC',
-      platform: 'Google Ads',
-      retention: '2 years',
-      wildcard: '0',
-    },
-    frameIdList: [3964, 0, 3967],
-    headerType: 'request',
-    isFirstParty: false,
-    parsedCookie: {
-      domain: '.google.com',
-      expires: '2024-08-16T09:35:25.402Z',
-      httponly: true,
-      name: '__Secure-3PSIDCC',
-      path: '/',
-      samesite: 'none',
-      secure: true,
-      value:
-        'APoG2W9JZc9seJQs4d_vORdcjDLDbTsEZC4lI68BgWKHrHlo7duSmRgyT1AbvVwq-ZwQoyVm-XU',
-    },
-    url: 'https://www.google.com/recaptcha/api2/anchor?ar=1&k=6Le6JpwaAAAAAFkASVap1OUThl-lQSJC0r9kLl2I&co=aHR0cHM6Ly9ydGNhbXAuY29tOjQ0Mw..&hl=en&v=QybaJej5brGL8d7EvWmfKMZU&theme=light&size=invisible&badge=bottomright&cb=hztbhk44za6a',
-    isCookieSet: true,
-  },
-  _fuid: {
-    analytics: {
-      category: '',
-      dataController: '',
-      description: '',
-      domain: '',
-      gdprUrl: '',
-      name: '',
-      platform: '',
-      retention: '',
-      wildcard: '',
-    },
-    frameIdList: [0],
-    headerType: 'request',
-    isFirstParty: true,
-    parsedCookie: {
-      domain: '.rtcamp.com',
-      expires: '2024-08-12T15:47:39.137Z',
-      httponly: false,
-      name: '_fuid',
-      path: '/',
-      samesite: '',
-      secure: false,
-      value: 'NjE0Zjc5MDAtMTM5NC00NjJmLWE5ODctZGI1Mzc3YWY0YWRh',
-    },
-    url: 'https://rtcamp.com/wp-content/mu-plugins/jetpack-12.4/css/jetpack.css?ver=12.4',
-    isCookieSet: true,
-  },
-  _ga: {
-    analytics: {
-      category: 'Analytics',
-      dataController: 'Google',
-      description: 'ID used to identify users',
-      domain:
-        "google-analytics.com (3rd party) or advertiser's website domain (1st party)",
-      gdprUrl: 'https://privacy.google.com/take-control.html',
-      name: '_ga',
-      platform: 'Google Analytics',
-      retention: '2 years',
-      wildcard: '0',
-    },
-    frameIdList: [0],
-    headerType: 'request',
-    isFirstParty: true,
-    parsedCookie: {
-      domain: '.rtcamp.com',
-      expires: '2024-09-20T09:35:20.538Z',
-      httponly: false,
-      name: '_ga',
-      path: '/',
-      samesite: '',
-      secure: false,
-      value: 'GA1.1.2118227284.1688917659',
-    },
-    url: 'https://rtcamp.com/wp-content/mu-plugins/jetpack-12.4/css/jetpack.css?ver=12.4',
-    isCookieSet: true,
-  },
-  _ga_7HKDVLRRV4: {
-    analytics: {
-      category: 'Analytics',
-      dataController: 'Google',
-      description: 'ID used to identify users',
-      domain:
-        "google-analytics.com (3rd party) or advertiser's website domain (1st party)",
-      gdprUrl: 'https://privacy.google.com/take-control.html',
-      name: '_ga_*',
-      platform: 'Google Analytics',
-      retention: '2 years',
-      wildcard: '1',
-    },
-    frameIdList: [0],
-    headerType: 'request',
-    isFirstParty: true,
-    parsedCookie: {
-      domain: '.rtcamp.com',
-      expires: '2024-09-20T09:35:24.713Z',
-      httponly: false,
-      name: '_ga_7HKDVLRRV4',
-      path: '/',
-      samesite: '',
-      secure: false,
-      value: 'GS1.1.1692264828.19.1.1692264924.55.0.0',
-    },
-    url: 'https://rtcamp.com/wp-content/mu-plugins/jetpack-12.4/css/jetpack.css?ver=12.4',
-    isCookieSet: true,
-  },
-  _parsely_session: {
-    analytics: {
-      category: '',
-      dataController: '',
-      description: '',
-      domain: '',
-      gdprUrl: '',
-      name: '',
-      platform: '',
-      retention: '',
-      wildcard: '',
-    },
-    frameIdList: [0],
-    headerType: 'request',
-    isFirstParty: true,
-    parsedCookie: {
-      domain: '.rtcamp.com',
-      expires: '2023-08-17T10:05:20.000Z',
-      httponly: false,
-      name: '_parsely_session',
-      path: '/',
-      samesite: '',
-      secure: false,
-      value:
-        '{%22sid%22:8%2C%22surl%22:%22https://rtcamp.com/%22%2C%22sref%22:%22%22%2C%22sts%22:1692264828254%2C%22slts%22:1692261902478}',
-    },
-    url: 'https://rtcamp.com/wp-content/mu-plugins/jetpack-12.4/css/jetpack.css?ver=12.4',
-    isCookieSet: false,
-  },
-  _parsely_visitor: {
-    analytics: {
-      category: '',
-      dataController: '',
-      description: '',
-      domain: '',
-      gdprUrl: '',
-      name: '',
-      platform: '',
-      retention: '',
-      wildcard: '',
-    },
-    frameIdList: [0],
-    headerType: 'request',
-    isFirstParty: true,
-    parsedCookie: {
-      domain: '.rtcamp.com',
-      expires: '2024-09-15T19:35:20.000Z',
-      httponly: false,
-      name: '_parsely_visitor',
-      path: '/',
-      samesite: '',
-      secure: false,
-      value:
-        '{%22id%22:%22pid=9ebac1f53e4598f174e600d465fc30b9%22%2C%22session_count%22:8%2C%22last_session_ts%22:1692264828254}',
-    },
-    url: 'https://rtcamp.com/wp-content/mu-plugins/jetpack-12.4/css/jetpack.css?ver=12.4',
-    isCookieSet: true,
-  },
-  track_uid: {
-    analytics: {
-      category: '',
-      dataController: '',
-      description: '',
-      domain: '',
-      gdprUrl: '',
-      name: '',
-      platform: '',
-      retention: '',
-      wildcard: '',
-    },
-    frameIdList: [0],
-    headerType: 'request',
-    isFirstParty: true,
-    parsedCookie: {
-      domain: '.rtcamp.com',
-      expires: '2024-07-11T03:07:19.000Z',
-      httponly: false,
-      name: 'track_uid',
-      path: '/',
-      samesite: '',
-      secure: false,
-      value: '2092510f-541d-4498-9d78-37829e5e5360',
-    },
-    url: 'https://rtcamp.com/wp-content/mu-plugins/jetpack-12.4/css/jetpack.css?ver=12.4',
-    isCookieSet: true,
-  },
-};
